抽取截屏库
